fix(changelog): address review feedback
