+++ .prettierignore
@@ -1,6 +1,7 @@
 # Generated — never hand-edit; regenerated by `npm run api` / sqlc / openapi-typescript
 frontend/src/api/schema.ts
 frontend/src/renderer/routeTree.gen.ts
+frontend/src/landing/data/changelog.json
 backend/internal/httpd/apispec/openapi.yaml
 
 # Build outputs
